requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot[all]==20.7
python-dotenv==1.0.0
openpyxl==3.1.2